Fix fuzzer setup
