+++ compiler/vm/benches/utils.rs
@@ -8,7 +8,7 @@ use candy_frontend::{
     mir_optimize::OptimizeMirStorage,
     module::{
         GetModuleContentQuery, InMemoryModuleProvider, Module, ModuleDbStorage, ModuleKind,
-        ModuleProvider, ModuleProviderOwner, MutableModuleProviderOwner, Package, PackagesPath,
+        ModuleProvider, ModuleProviderOwner, MutableModuleProviderOwner, Package,
     },
     position::PositionConversionStorage,
     rcst_to_cst::RcstToCstStorage,
@@ -20,13 +20,12 @@ use candy_vm::{
     lir::Lir,
     mir_to_lir::compile_lir,
     tracer::DummyTracer,
-    Vm,
+    PopulateInMemoryProviderFromFileSystem, Vm,
 };
 use lazy_static::lazy_static;
 use rustc_hash::FxHashMap;
-use std::{borrow::Borrow, fs};
+use std::borrow::Borrow;
 use tracing::warn;
-use walkdir::WalkDir;
 
 const TRACING: TracingConfig = TracingConfig::off();
 lazy_static! {
@@ -78,8 +77,8 @@ pub fn setup_and_compile(source_code: &str) -> Lir {
 
 pub fn setup() -> Database {
     let mut db = Database::default();
-    load_package("Builtins", &mut db.module_provider);
-    load_package("Core", &mut db.module_provider);
+    db.module_provider.load_package_from_file_system("Builtins");
+    db.module_provider.load_package_from_file_system("Core");
     db.module_provider.add_str(&MODULE, r#"_ = use "Core""#);
 
     // Load `Core` into the cache.
@@ -92,30 +91,6 @@ pub fn setup() -> Database {
     }
 
     db
-}
-fn load_package(package_name: impl Into<String>, module_provider: &mut InMemoryModuleProvider) {
-    let package_name = package_name.into();
-    let packages_path = PackagesPath::try_from("../../packages").unwrap();
-    let package_path = packages_path.join(package_name.clone());
-    let package = Package::Managed(package_name.into());
-
-    for file in WalkDir::new(&package_path)
-        .into_iter()
-        .map(|it| it.unwrap())
-        .filter(|it| it.file_type().is_file())
-        .filter(|it| it.file_name().to_string_lossy().ends_with(".candy"))
-    {
-        let module = Module::from_package_and_path(
-            &packages_path,
-            package.clone(),
-            file.path(),
-            ModuleKind::Code,
-        )
-        .unwrap();
-
-        let source_code = fs::read_to_string(file.path()).unwrap();
-        module_provider.add_str(&module, source_code);
-    }
 }
 
 pub fn compile(db: &mut Database, source_code: &str) -> Lir {
